+++ src/main.ts
@@ -28,6 +28,10 @@ export async function main() {
             type: "boolean",
             describe: "Delete content types as well",
             default: false
+        }).option("assets", {
+            type: "boolean",
+            describe: "Delete assets as well",
+            default: false
         }).option("yes", {
             type: "boolean",
             describe: "Auto-confirm delete prompt",
@@ -44,6 +48,7 @@ export async function main() {
     const verbose: boolean = argv["verbose"];
     const batchSize: number = argv["batch-size"];
     const isContentTypes: boolean = argv["content-types"];
+    const isAssets: boolean = argv["assets"];
     const yes: boolean = argv["yes"];
 
     const env: string = argv["env"] || 'master';
@@ -68,6 +73,14 @@ export async function main() {
         }
         await deleteContentTypes(contentfulSpace, batchSize, verbose, env);
     }
+
+    if (isAssets) {
+        if (!yes) {
+            if (!await promptForAssetsConfirmation(spaceId, env))
+                return;
+        }
+        await deleteAssets(contentfulSpace, batchSize, verbose, env);
+    }
 }
 
 async function promptForEntriesConfirmation(spaceId: string, environment: string) {
@@ -86,6 +99,42 @@ async function promptForContentTypesConfirmation(spaceId: string, environment: s
         message: `Do you really want to delete all content types from space ${spaceId}:${environment}?`
     }]);
     return a.yes;
+}
+
+async function promptForAssetsConfirmation(spaceId: string, environment: string) {
+    const a: any = await inquirer.prompt([{
+        type: "confirm",
+        name: "yes",
+        message: `Do you really want to delete all assets/media from space ${spaceId}:${environment}?`
+    }]);
+    return a.yes;
+}
+
+async function deleteAssets(contentfulSpace: any, batchSize: number, verbose: boolean,  environment: string) {
+    const selectedEnvironment = await contentfulSpace.getEnvironment(environment);
+    const entriesMetadata = await selectedEnvironment.getAssets({
+        include: 0,
+        limit: 0
+    });
+    let totalEntries = entriesMetadata.total;
+    console.log(`Deleting ${totalEntries} entries`);
+
+    // tslint:disable-next-line:max-line-length
+    const entriesProgressBar = new ProgressBar("Deleting entries [:bar], rate: :rate/s, done: :percent, time left: :etas", { total: totalEntries });
+    do {
+        const assets = await selectedEnvironment.getAssets({
+            include: 0,
+            limit: batchSize
+        });
+        totalEntries = assets.total;
+
+        const promises: Array<Promise<void>> = [];
+        for (const asset of assets.items) {
+            const promise = unpublishAndDeleteEntry(asset, entriesProgressBar, verbose);
+            promises.push(promise);
+        }
+        await Promise.all(promises);
+    } while (totalEntries > batchSize);
 }
 
 async function deleteEntries(contentfulSpace: any, batchSize: number, verbose: boolean,  environment: string) {
@@ -115,16 +164,16 @@ async function deleteEntries(contentfulSpace: any, batchSize: number, verbose: b
     } while (totalEntries > batchSize);
 }
 
-async function unpublishAndDeleteEntry(entry: any, progressBar: ProgressBar, verbose: boolean) {
+async function unpublishAndDeleteEntry(assetOrEntry: any, progressBar: ProgressBar, verbose: boolean) {
     try {
-        if (entry.isPublished()) {
+        if (assetOrEntry.isPublished()) {
             if (verbose)
-                console.log(`Unpublishing entry "${entry.sys.id}"`);
-            await entry.unpublish();
+                console.log(`Unpublishing entry "${assetOrEntry.sys.id}"`);
+            await assetOrEntry.unpublish();
         }
         if (verbose)
-            console.log(`Deleting entry '${entry.sys.id}"`);
-        await entry.delete();
+            console.log(`Deleting entry '${assetOrEntry.sys.id}"`);
+        await assetOrEntry.delete();
     } catch (e) {
         console.log(e);
         // Continue if something went wrong with Contentful